Fiddling with format still
I don't want to store redundant information w.r.t. length, variant id, span, etc. so I've taken them out of the individual records and put them into an array explicitly.

There is also the issue that the gzipped 2bit data bytes are holding multiple variants in one AltData entry.  This means that multiple entries in the VariantIndex array will point to a single entry (in the AltData record(s)).  The VariantType will hold which of the (now two) structures to consult (either the AltData or the AltOverflowRec).  VariantIds can be in order (maybe we should have a flag for that) but can be jumbled around if need be.