VariantIndex-AltData elaboration
VariantIndex has a confusing idea of having multiple variants point to the same `AltData` record.  The `AltData` record as of now is the gzip of a twobit file, so can contain multiple records and that's why it's set up like this.  A note has been added in the `Notes` section to indicate this behavior.