Changin some sizes and names
We don't need to be clever for some lengths and offsets, just make them 8byte and don't worry about it.  "Len" was repeated in a structure.  One is the length of the record, the other is the length of the run of low quality 'no-calls'.  The latter is named 'LoqLen' appropriately.