Code array of bytes instead of raw dlug
First off, the Code needed to be an array in the FinalOverflowMap.  Second off we don't need to overcomplicate things.  It can be a simple byte array that has a single byte code for the type of record.  If we really need to extend past 256 options we can reserve a byte to indicate the data bytes themselves hold the information about what type of overflow record it is.