clarifying what Offset and Position are
There's ambiguity whether `Offset` and `Position` refer to the position of the tile step or the entry in `LoqInfo`.  Comments have been updated to clarify and `Position` has been named to `StepPosition`.

Entry `k` in `Offset` (e.g. `Offset[k]`) refers to the the `Stride*k` low quality entries byte offset, starting from `LoqInfo[0]`.  For example, if `Stride` were 256 then `Offset[2]` (if it exists) has the byte offset, using `LoqInfo[0]` as the reference point`, of entry `LoqInfo[768]`.

Entry `k` in `StepPosition` (e.g. `StepPosition[k]`) refers to the `Stride*k` low quality step position that `LoqInfo[Stride*k]`refers to.  For example, if stride were 256, then `Offset[3]` (if it exists) refers to the tile step that `LoqInfo[768]` refers to.